% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_climate_timeseries.R
\name{fits_from_params}
\alias{fits_from_params}
\title{Fit a spline to a collection of timeseries, with a possible roughness penalty}
\usage{
fits_from_params(
  data_to_smooth,
  timepoints,
  nbasis_in,
  lambda_in = 0,
  basis_type = "fourier",
  norder_in = 4
)
}
\arguments{
\item{data_to_smooth}{A data frame. Rows correspond to different time points
and columns correspond to different variables.}

\item{timepoints}{A vector (numeric). Corresponds to the measurement
timepoints of each variable.}

\item{nbasis_in}{Integer. The number of basis functions to use.}

\item{lambda_in}{Float. Defaults to 0. The value of the smoothing parameter,
lambda.}

\item{basis_type}{String. Defaults to "fourier". Can be either "fourier" or
"bspline" to select different families of basis functions.}

\item{norder_in}{Integer. Defaults to 4. Only relevant if basis_type =
TRUE. Controls the order of the bspline functions.}
}
\value{
An fdSmooth object.
}
\description{
Fit a spline to a collection of timeseries, with a possible roughness penalty
}
